requirements.txt
astroid==2.2.5
awscli==1.16.238
awsebcli==3.15.3
blessed==1.15.0
botocore==1.12.228
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
defusedxml==0.6.0
Django==2.1.1
django-allauth==0.40.0
django-cleanup==4.0.0
django-extensions==2.2.1
django-jet==1.0.8
django-widget-tweaks==1.4.5
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
isort==4.3.21
jmespath==0.9.4
jsonschema==2.6.0
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.5.9
Pillow==6.1.0
pyasn1==0.4.7
pylint==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.2
